refactor: DiaryTextView 구현
